Undress first
When you use the 'try on' button, it will now first Undress your model.
That way if for example you had a tabard on, but the person you're
basing your outfit on doesn't, it won't continue to show your tabard.

Also, all the dressing room now has an 'Undress' button.  This is there
if you have ever tried to look at a chest piece but couldn't because you
had a tabard on.  Or couldn't see your new boots because of your robe.
Or maybe bracers will matter the look of now if your gloves weren't
covering them up ;-)