Legacy and Class specific mount support
Instead of "not mounted" message coming up when looking at legacy mounts
(like the Ancient Frostsaber or Black Proto-Drake) or class mounts (like
the Charger or Felsteed) a popup dialog with some details will be given
instead.

Please report any mounts which are not caught by this update but I think
this will handle them.